Refactor leap Set/node to long instead of double etc..

+++ src/leapListReg/LeapNode.java
@@ -5,20 +5,21 @@ import utils.Trie;
 
 public class LeapNode {
 	boolean live = true;
-	double low;
-	double high;
+	long low;
+	long high;
 	int count;
-	int level;
+	byte level;
 	LeapSet [] data = new LeapSet[LeapList.NODE_SIZE]; //the array must be sorted by keys so that LeapSet with the smallest key is at LeapSet[0] etc.
 	LeapNode [] next = new LeapNode [LeapList.MAX_LEVEL];
 	Trie trie;
 	
-	public LeapNode (boolean live, double low, double high, int count, int level, LeapSet[] sortedPairs){
+	public LeapNode (boolean live, long low, long high, int count, byte level, LeapSet[] sortedPairs){
 		if (sortedPairs != null)
 			trie = new Trie(sortedPairs, sortedPairs.length);
 	}
 	
 	public LeapNode (){
+		// TODO: init new trie  
 	}
 
 }
